Rewrite lender loans and offers optimistics
