+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveLoansTable/ActiveLoansTable.module.less
