+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveLoansTable/TableCells/APRCell.tsx
