+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveLoansTable/TableCells/ActionsCell.tsx
@@ -5,6 +5,8 @@ import { isEmpty } from 'lodash'
 import { Button } from '@banx/components/Buttons'
 
 import { Loan, Offer } from '@banx/api/core'
+import { useHiddenNftsMints } from '@banx/pages/BorrowPage/hooks'
+import { OfferOptimistic } from '@banx/store'
 import {
   isLoanActiveOrRefinanced,
   isLoanLiquidated,
@@ -12,7 +14,6 @@ import {
   trackPageEvent,
 } from '@banx/utils'
 
-import { useHiddenNftsMints, useOptimisticOffers } from '../../../hooks'
 import { findBestOffer } from '../helpers'
 import { useLendLoansTransactions } from '../hooks'
 
@@ -21,8 +22,9 @@ import styles from '../ActiveLoansTable.module.less'
 interface ActionsCellProps {
   loan: Loan
   offers: Record<string, Offer[]>
-  updateOrAddOffer: (offer: Offer) => void
+  updateOrAddOffer: (offer: Offer[]) => void
   updateOrAddLoan: (loan: Loan) => void
+  optimisticOffers: OfferOptimistic[]
 }
 
 export const ActionsCell: FC<ActionsCellProps> = ({
@@ -30,9 +32,9 @@ export const ActionsCell: FC<ActionsCellProps> = ({
   offers,
   updateOrAddOffer,
   updateOrAddLoan,
+  optimisticOffers,
 }) => {
-  const { offers: optimisticOffers } = useOptimisticOffers()
-  const { addMints } = useHiddenNftsMints()
+  const { add: addMints } = useHiddenNftsMints()
 
   const bestOffer = useMemo(() => {
     return findBestOffer({ loan, offers, optimisticOffers })
