+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveLoansTable/TableCells/InterestCell.tsx
