+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveLoansTable/TableCells/StatusCell.tsx
