+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveLoansTable/TableCells/index.ts
