+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveLoansTable/columns.tsx
@@ -2,6 +2,7 @@ import { ColumnType } from '@banx/components/Table'
 import { HeaderCell, NftInfoCell, createSolValueJSX } from '@banx/components/TableComponents'
 
 import { Loan, Offer } from '@banx/api/core'
+import { OfferOptimistic } from '@banx/store'
 import { formatDecimal } from '@banx/utils'
 
 import { calculateLentValue } from '../OfferCard/helpers'
@@ -9,11 +10,17 @@ import { APRCell, ActionsCell, InterestCell, StatusCell } from './TableCells'
 
 interface GetTableColumns {
   offers: Record<string, Offer[]>
-  updateOrAddOffer: (offer: Offer) => void
+  updateOrAddOffer: (offer: Offer[]) => void
   updateOrAddLoan: (loan: Loan) => void
+  optimisticOffers: OfferOptimistic[]
 }
 
-export const getTableColumns = ({ offers, updateOrAddOffer, updateOrAddLoan }: GetTableColumns) => {
+export const getTableColumns = ({
+  offers,
+  updateOrAddOffer,
+  updateOrAddLoan,
+  optimisticOffers,
+}: GetTableColumns) => {
   const columns: ColumnType<Loan>[] = [
     {
       key: 'collateral',
@@ -72,6 +79,7 @@ export const getTableColumns = ({ offers, updateOrAddOffer, updateOrAddLoan }: G
           offers={offers}
           updateOrAddOffer={updateOrAddOffer}
           updateOrAddLoan={updateOrAddLoan}
+          optimisticOffers={optimisticOffers}
           loan={loan}
         />
       ),
