+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveLoansTable/index.ts
