+++ src/pages/OffersPage/components/OffersTabContent/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from './useLenderLoansAndOffers'
 export * from './useOffersTabContent'
